Fix for "authorizations for these names not found or expired" on Certificate Renewal
Due to "authorizations for these names not found or expired" error we now create an Alias every time (for both - New Bindings and Renewals). The reason for this is that the authorization for the domain is active for one month but the certificates are for 3 months, so alot of users started getting this error after trying to renew their certificate after three months.